feat(serve): Adding `npm run serve` command
`npm run serve` will build the current instantsearch and expose the
file on http://127.0.0.1:8080/instantsearch.js. It will get
automatically updated in each change.

This is useful when you want to test your branch in a specific
project. Just run `npm run serve` in the instantsearch repo and link
to the file in your project.

I've regenerated the shrinkwrap but had to run a `npm prune` before
shrinkwrap allowed me to do it, so I updated the doc accordingly.